Refactor Layer to include CustomLayer interface functions
