+++ src/interactivity/Interactivity.js
@@ -1,5 +1,5 @@
 import mitt from 'mitt';
-import Layer from '../layers/Layer';
+import Layer from '../Layer';
 import { WM_R, projectToWebMercator } from '../utils/util';
 import { wToR } from '../client/rsys';
